Updated version number (to 0.04)
Version 0.04 is a major milestone, reverting the changes in 0.03 to take
us back to 0.02, but with connotations of progress and accomplishment.

Do not let go of the fear, though, that there are only so many numbers.
To exhaust them like this is wasteful, and is only passable by necessity
of version number linearity. Though, I'm not quite sure this is the
source of my discomfort.