Deleted mention of when P(G <= np) sits at borders
For p = k/n and k = {0, 1, ..., n},
P(G(U, n)) will either consistently sit at the lower or upper bound,
depending on the interpretation of whether the 0-percentile includes
indexes of 0. This confuses me, so I don't want to talk about it inline,
but I might make brief mention in the readme that is currently under
construction.